+++ src/actor/reactor.rs
@@ -2110,6 +2110,7 @@ impl Reactor {
                     && let Some(wid) = tracked_window
                 {
                     self.correct_straggler_after_return(wid, sid);
+                    self.keep_record_window_off_an_unknown_display(wid, sid);
                     self.note_window_appeared_while_away(wid, sid);
                 }
                 return topology_workflow::handle_window_server_appeared(